move get_file_type into file_utils module
