+++ src/packs.rs
@@ -10,6 +10,7 @@ use std::path::Path;
 use std::path::PathBuf;
 pub(crate) mod checker;
 pub mod cli;
+pub(crate) mod file_utils;
 mod inflector_shim;
 pub mod logger;
 mod pack_set;
