+++ src/packs/parser.rs
@@ -9,7 +9,7 @@ pub(crate) use ruby::packwerk::extractor::process_from_path as process_from_ruby
 pub(crate) mod erb;
 pub(crate) use erb::packwerk::extractor::process_from_path as process_from_erb_path;
 
-use super::ProcessedFile;
+use super::{file_utils::get_file_type, ProcessedFile};
 
 #[derive(PartialEq, Debug)]
 pub enum SupportedFileType {
@@ -51,33 +51,10 @@ pub fn process_files_with_cache<T: Cache + Send + Sync>(
         .collect()
 }
 
-fn get_file_type(path: &Path) -> Option<SupportedFileType> {
-    let ruby_special_files = vec!["Gemfile", "Rakefile"];
-    let ruby_extensions = vec!["rb", "rake", "builder", "gemspec", "ru"];
-
-    let extension = path.extension();
-    // Eventually, we can have packs::parser::ruby, packs::parser::erb, etc.
-    // These would implement a packs::parser::interface::Parser trait and can
-    // hold the logic for determining if a parser can parse a file.
-
-    let is_ruby_file = ruby_extensions
-        .into_iter()
-        .any(|ext| extension.map_or(false, |e| e == ext))
-        || ruby_special_files.iter().any(|file| path.ends_with(file));
-
-    let is_erb_file = path.extension().map_or(false, |ext| ext == "erb");
-
-    if is_ruby_file {
-        Some(SupportedFileType::Ruby)
-    } else if is_erb_file {
-        Some(SupportedFileType::Erb)
-    } else {
-        None
-    }
-}
-
 #[cfg(test)]
 mod tests {
+    use crate::packs::file_utils::get_file_type;
+
     use super::*;
 
     fn assert_is_ruby(filename: &str) {
